Better exhausted playlist handling
The thing was that when remote playlist comes to the end, Vlc stops
playing. When new track is uploaded, it is added to remote playlist,
then our Remote Control sees that Vlc does not play and tells him to
start. And Vlc starts to play from the very beginning of the playlist.

Since we do not need to store that remote playlist, we just clean it up
before adding new track, if it is stopped.